# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Laravel with PostgreSQL - Created Muhamad Ferdinal",
        "version": "1.0.0"
    },
    "paths": {
...
        "/produk-stock": {
            "get": {
                "tags": [
                    "Product Stock"
                ],
                "summary": "Get list of product stocks",
                "description": "Returns list of product stocks",
                "operationId": "findAllProductStock",
                "responses": {
                    "200": {
                        "description": "Pengambilan Data Produk Stok Berhasil!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Pengambilan Data Produk Stok!"
                    }
                }
            }
        },
        "/produk-stock/{produk_id}": {
            "get": {
                "tags": [
                    "Product Stock"
                ],
                "summary": "Get single product stock data",
                "description": "Returns single product stock data",
                "operationId": "findOneProductStock",
                "parameters": [
                    {
                        "name": "produk_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengambilan Data Produk Stok Berhasil!"
                    },
                    "404": {
                        "description": "Data Produk Stok Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Pengambilan Data Produk Stok!"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product Stock"
                ],
                "summary": "Update product stock data",
                "description": "Update product stock data",
                "operationId": "updateOneProductStock",
                "parameters": [
                    {
                        "name": "produk_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "stok"
                                ],
                                "properties": {
                                    "stok": {
                                        "description": "Stok Produk (Required)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Produk Stok Berhasil Diperbarui!"
                    },
                    "404": {
                        "description": "Data Produk Stok Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Memperbarui Data Produk Stok!"
                    }
                }
            }
        },
...
    }
}